Añade notas y horarios personalizados en asistencia
- Se agrega columna `notes` a `assistance_incident` para comentarios adicionales.
- Se añade columna `source_check_in` a `employee_check_ins` para registrar la hora original de entrada.
- Se crean tablas `activity_work_schedule_configs` y `employee_work_schedule_assignment` para asignación flexible de horarios por actividad y empleado.
- Se actualizan snapshots y modelos de EF Core.
- Corrección menor en el tipo de parámetro de empresa en `AttendanceRecordsService`.

+++ PrenominaApi/Services/AttendanceRecordsService.cs
@@ -174,7 +174,7 @@ namespace PrenominaApi.Services
             var paginatedCodesInt = paginatedEmployeeCodes.Select(c => (int)c).ToList();
             var schedulesByEmployee = _scheduleResolver.GetSchedulesForEmployees(
                 paginatedCodesInt,
-                filter.Company,
+                (int)filter.Company,
                 periodDates.StartDate,
                 periodDates.ClosingDate);
